I've added a lot more structure to the file. I've defined several new functions for use in our code.
	introduction()
	are_all_ships_dead(player)
	take_a_shot()
	coin_flip()

There is also a debug switch at the top of the file that controls the output of a bunch of tests at the end of the file.

I'm having trouble getting recursion to return the right value in the take_a_shot() function, and I may have to resort to a while loop.